Support parameterization using tags and env vars
buildxce now recognizes a parameters cell, if present, in the
notebook. A parameters cell is denoted by the same convention
used in papermill: a "parameters" tag applied by the user.
If such a cell is present, the values of any variables defined
in it can be overridden by environment variables at runtime.