Implement basic EO Application Package support
xcetool now supports an --eoap option to the build subcommand, which
produces a basic, runnable EO Application Package as a CWL file along
with the Docker image. Input parameters are supported (they can be
supplied to the CWL runner and will be passed through to the user
code in the container).

However, EOAP stage-in and stage-out is not yet supported, so actual
datasets cannot yet be passed to and from the Application Package.